rename OhsomeQgisDialogUI to OhsomeToolsDialogUI
